fix and add tests
